Ny komponent behandlingsvalgRadioGroup.tsx
Inneholder både radioGroup og inputfelt dersom bruker velger "Jeg skal oppgi endelig avgift selv"
